+++ src/test/java/com/example/xbankbackend/services/TransactionsServiceTest.java
@@ -26,6 +26,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 import java.util.UUID;
@@ -56,7 +57,7 @@ public class TransactionsServiceTest {
         transaction.setTransactionType(TransactionType.PAYMENT);
         transaction.setReceiverId(receiverId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         assertThrows(IllegalArgumentException.class, () -> transactionsService.deposit(transaction));
     }
@@ -69,7 +70,7 @@ public class TransactionsServiceTest {
         transaction.setTransactionType(TransactionType.DEPOSIT);
         transaction.setReceiverId(receiverId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         assertThrows(IllegalArgumentException.class, () -> transactionsService.deposit(transaction));
     }
@@ -83,7 +84,7 @@ public class TransactionsServiceTest {
         transaction.setReceiverId(receiverId);
         transaction.setSenderId(UUID.randomUUID());
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         assertThrows(IllegalArgumentException.class, () -> transactionsService.deposit(transaction));
     }
@@ -96,7 +97,7 @@ public class TransactionsServiceTest {
         transaction.setTransactionType(TransactionType.DEPOSIT);
         transaction.setReceiverId(receiverId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         when(bankAccountRepository.exists(receiverId)).thenReturn(false);
 
@@ -111,7 +112,7 @@ public class TransactionsServiceTest {
         transaction.setTransactionType(TransactionType.DEPOSIT);
         transaction.setReceiverId(receiverId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(-5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(-5000.0));
 
         when(bankAccountRepository.exists(receiverId)).thenReturn(true);
 
@@ -126,15 +127,15 @@ public class TransactionsServiceTest {
         transaction.setTransactionType(TransactionType.DEPOSIT);
         transaction.setReceiverId(receiverId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         when(bankAccountRepository.getCurrency(receiverId)).thenReturn(CurrencyType.RUB);
         when(bankAccountRepository.exists(receiverId)).thenReturn(true);
-        when(currencyRateService.convert(any(), any(), any())).thenReturn(5000.0f);
+        when(currencyRateService.convert(any(), any(), any())).thenReturn(BigDecimal.valueOf(5000.0));
 
         transactionsService.deposit(transaction);
 
-        verify(bankAccountRepository).increaseBalance(receiverId, 5000.0f);
+        verify(bankAccountRepository).increaseBalance(receiverId, BigDecimal.valueOf(5000.0));
     }
 
     // validateTransfer
@@ -148,7 +149,7 @@ public class TransactionsServiceTest {
         transaction.setReceiverId(receiverId);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         assertThrows(IllegalArgumentException.class, () -> transactionsService.transfer(transaction));
     }
@@ -163,7 +164,7 @@ public class TransactionsServiceTest {
         transaction.setReceiverId(receiverId);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         assertThrows(IllegalArgumentException.class, () -> transactionsService.transfer(transaction));
     }
@@ -178,7 +179,7 @@ public class TransactionsServiceTest {
         transaction.setReceiverId(receiverId);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         assertThrows(IllegalArgumentException.class, () -> transactionsService.transfer(transaction));
     }
@@ -193,7 +194,7 @@ public class TransactionsServiceTest {
         transaction.setReceiverId(receiverId);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         when(bankAccountRepository.exists(receiverId)).thenReturn(false);
 
@@ -210,7 +211,7 @@ public class TransactionsServiceTest {
         transaction.setReceiverId(receiverId);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         when(bankAccountRepository.exists(receiverId)).thenReturn(true);
         when(bankAccountRepository.exists(senderId)).thenReturn(false);
@@ -228,11 +229,11 @@ public class TransactionsServiceTest {
         transaction.setReceiverId(receiverId);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         when(bankAccountRepository.exists(receiverId)).thenReturn(true);
         when(bankAccountRepository.exists(senderId)).thenReturn(true);
-        when(bankAccountRepository.getBalance(senderId)).thenReturn(4000.0f);
+        when(bankAccountRepository.getBalance(senderId)).thenReturn(BigDecimal.valueOf(4000.0));
 
         assertThrows(InsufficientFundsException.class, () -> transactionsService.transfer(transaction));
     }
@@ -247,11 +248,11 @@ public class TransactionsServiceTest {
         transaction.setReceiverId(receiverId);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(-5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(-5000.0));
 
         when(bankAccountRepository.exists(receiverId)).thenReturn(true);
         when(bankAccountRepository.exists(senderId)).thenReturn(true);
-        when(bankAccountRepository.getBalance(senderId)).thenReturn(6000.0f);
+        when(bankAccountRepository.getBalance(senderId)).thenReturn(BigDecimal.valueOf(6000.0));
 
         assertThrows(IllegalArgumentException.class, () -> transactionsService.transfer(transaction));
     }
@@ -266,18 +267,18 @@ public class TransactionsServiceTest {
         transaction.setReceiverId(receiverId);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         when(bankAccountRepository.exists(receiverId)).thenReturn(true);
         when(bankAccountRepository.exists(senderId)).thenReturn(true);
-        when(bankAccountRepository.getBalance(senderId)).thenReturn(6000.0f);
+        when(bankAccountRepository.getBalance(senderId)).thenReturn(BigDecimal.valueOf(6000.0));
         when(bankAccountRepository.getCurrency(receiverId)).thenReturn(CurrencyType.RUB);
-        when(currencyRateService.convert(any(), any(), any())).thenReturn(5000.0f);
+        when(currencyRateService.convert(any(), any(), any())).thenReturn(BigDecimal.valueOf(5000.0));
 
         transactionsService.transfer(transaction);
 
-        verify(bankAccountRepository).increaseBalance(receiverId, 5000.0f);
-        verify(bankAccountRepository).decreaseBalance(senderId, 5000.0f);
+        verify(bankAccountRepository).increaseBalance(receiverId, BigDecimal.valueOf(5000.0));
+        verify(bankAccountRepository).decreaseBalance(senderId, BigDecimal.valueOf(5000.0));
     }
 
     // validatePayment
@@ -289,7 +290,7 @@ public class TransactionsServiceTest {
         transaction.setTransactionType(TransactionType.DEPOSIT);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         assertThrows(IllegalArgumentException.class, () -> transactionsService.pay(transaction));
     }
@@ -302,7 +303,7 @@ public class TransactionsServiceTest {
         transaction.setTransactionType(TransactionType.PAYMENT);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         assertThrows(IllegalArgumentException.class, () -> transactionsService.pay(transaction));
     }
@@ -316,7 +317,7 @@ public class TransactionsServiceTest {
         transaction.setReceiverId(UUID.randomUUID());
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         assertThrows(IllegalArgumentException.class, () -> transactionsService.pay(transaction));
     }
@@ -329,7 +330,7 @@ public class TransactionsServiceTest {
         transaction.setTransactionType(TransactionType.PAYMENT);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         when(bankAccountRepository.exists(senderId)).thenReturn(false);
 
@@ -344,10 +345,10 @@ public class TransactionsServiceTest {
         transaction.setTransactionType(TransactionType.PAYMENT);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         when(bankAccountRepository.exists(senderId)).thenReturn(true);
-        when(bankAccountRepository.getBalance(senderId)).thenReturn(4000.0f);
+        when(bankAccountRepository.getBalance(senderId)).thenReturn(BigDecimal.valueOf(4000.0));
 
         assertThrows(InsufficientFundsException.class, () -> transactionsService.pay(transaction));
     }
@@ -360,10 +361,10 @@ public class TransactionsServiceTest {
         transaction.setTransactionType(TransactionType.PAYMENT);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(-5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(-5000.0));
 
         when(bankAccountRepository.exists(senderId)).thenReturn(true);
-        when(bankAccountRepository.getBalance(senderId)).thenReturn(6000.0f);
+        when(bankAccountRepository.getBalance(senderId)).thenReturn(BigDecimal.valueOf(6000.0));
 
         assertThrows(IllegalArgumentException.class, () -> transactionsService.pay(transaction));
     }
@@ -376,14 +377,14 @@ public class TransactionsServiceTest {
         transaction.setTransactionType(TransactionType.PAYMENT);
         transaction.setSenderId(senderId);
         transaction.setCurrency(CurrencyType.RUB);
-        transaction.setAmount(5000.0f);
+        transaction.setAmount(BigDecimal.valueOf(5000.0));
 
         when(bankAccountRepository.exists(senderId)).thenReturn(true);
-        when(bankAccountRepository.getBalance(senderId)).thenReturn(6000.0f);
+        when(bankAccountRepository.getBalance(senderId)).thenReturn(BigDecimal.valueOf(6000.0));
 
         transactionsService.pay(transaction);
 
-        verify(bankAccountRepository).decreaseBalance(senderId, 5000.0f);
+        verify(bankAccountRepository).decreaseBalance(senderId, BigDecimal.valueOf(5000.0));
     }
 
     // getRecent
@@ -438,25 +439,25 @@ public class TransactionsServiceTest {
         Transaction transaction1 = new Transaction();
         transaction1.setTransactionType(TransactionType.DEPOSIT);
         transaction1.setCurrency(CurrencyType.RUB);
-        transaction1.setAmount(5000.0f);
+        transaction1.setAmount(BigDecimal.valueOf(5000.0));
         transaction1.setReceiverId(accountId);
 
         Transaction transaction2 = new Transaction();
         transaction2.setTransactionType(TransactionType.PAYMENT);
         transaction2.setCurrency(CurrencyType.RUB);
-        transaction2.setAmount(200.0f);
+        transaction2.setAmount(BigDecimal.valueOf(200.0));
         transaction2.setSenderId(accountId);
 
         TransactionResponse transactionResponse1 = new TransactionResponse();
         transactionResponse1.setTransactionType(TransactionType.DEPOSIT);
         transactionResponse1.setCurrency(CurrencyType.RUB);
-        transactionResponse1.setAmount(5000.0f);
+        transactionResponse1.setAmount(BigDecimal.valueOf(5000.0));
         transactionResponse1.setReceiverName("Test");
 
         TransactionResponse transactionResponse2 = new TransactionResponse();
         transactionResponse2.setTransactionType(TransactionType.PAYMENT);
         transactionResponse2.setCurrency(CurrencyType.RUB);
-        transactionResponse2.setAmount(200.0f);
+        transactionResponse2.setAmount(BigDecimal.valueOf(200.0));
         transactionResponse2.setSenderName("Test");
 
         List<Transaction> transactions = List.of(transaction1, transaction2);
